Revert "refactor: 좋아요페이지 리팩토링"
