+++ README.md
@@ -776,7 +776,6 @@
     - Error페이지 스타일링
 
 - 곽승헌
-
   - fix : 로그인 여부에 따라서 모달창 표시, 불필요 주석 제거
   - fix: 불필요 모달 창, 함수 제거
   - fix: 불필요 주석 삭제
@@ -805,20 +804,6 @@
 ### 9월 20일
 
 - 곽승헌
-
   - feat : 추천 코멘트 받는 함수 생성
   - fix: weatherState 기본 값 추가
   - fix: 날씨 정보 로딩이 끝날 때까지 Product 컴포넌트 대기
-
-- 황선우
-  - fix: 좋아요페이지 레이아웃에 맞도록 앨범스타일 및 마크업 조정
-  - fix : index를 위한 export 선언 조정
-  - fix: export 형태 조정에 따른 import 조정
-  - feat: 캘린더페이지 및 좋아요페이지 로그인 모달 삽입 및 좋아요페이지 수정
-    - 좋아요페이지 로그인 모달 추가 및 로그인 확인 로직 추가
-    - 날씨 필터 버튼 클릭 시 temperature 업데이트
-    - 필터링 로직 수정
-    - effect 훅을 통한 로그인 확인 로직 구현
-    - 날씨필터버튼에 마우스휠 이동 가능하도록 수정
-    - 버튼 컴포넌트 props 조정에 따른 수정
-  - style: CCM의 타입에 따라 cardWrapper 스타일 구분
