+++ package.json
@@ -19,7 +19,6 @@
     "@vitejs/plugin-react": "^4.3.1",
     "axios": "^1.7.7",
     "clsx": "^2.1.1",
-    "framer-motion": "^11.5.5",
     "kind-tiger": "^1.0.1",
     "motion": "^10.18.0",
     "pocketbase": "^0.21.4",
@@ -31,7 +30,6 @@
     "react-icons": "5.3.0",
     "react-lazyload": "^3.2.1",
     "react-router-dom": "^6.26.1",
-    "react-transition-group": "^4.4.5",
     "styled-components": "^6.1.13",
     "swiper": "^11.1.12",
     "use-immer": "0.10.0",
