+++ src/pages/CalendarPage.jsx
@@ -1,14 +1,13 @@
-import pb from '@/api/pocketbase';
-import { Bookmark, CommonModal } from '@/components';
-import { useUserStore } from '@/stores';
-import S from '@/styles/pages/Calendar.module.scss';
-import { useEffect, useState } from 'react';
+import { useState, useEffect } from 'react';
 import { Calendar } from 'react-calendar';
+import { Bookmark, CommonModal } from '@/components';
+import pb from '@/api/pocketbase';
+import S from '@/styles/pages/Calendar.module.scss';
 import 'react-calendar/dist/Calendar.css';
+import styled from 'styled-components';
 import { FaBookmark } from 'react-icons/fa';
 import { IoIosArrowBack, IoIosArrowForward } from 'react-icons/io';
-import styled from 'styled-components';
-
+import useUserStore from '@/stores/userStore';
 
 // 캘린더 스타일 커스터마이징
 const StyledCalendar = styled(Calendar)`
@@ -81,9 +80,6 @@ const isSameDay = (date1, date2) => {
 };
 
 const CalendarPage = () => {
-  const { isLoggedIn, initUser, user } = useUserStore();
-  const [isModalOpen, setIsModalOpen] = useState(true);
-
   const [date, setDate] = useState(new Date());
   const [visible, setVisible] = useState(false);
 
@@ -96,30 +92,15 @@ const CalendarPage = () => {
 
   const [calendarCollapsed, setCalendarCollapsed] = useState(false);
 
+  const [isModalOpen, setIsModalOpen] = useState(true);
 
-  useEffect(() => {
-    initUser();
-  }, [initUser]);
-
-  useEffect(() => {
-    if (!isLoggedIn) {
-      setIsModalOpen(true);
-    } else {
-      setIsModalOpen(false);
-    }
-  }, [isLoggedIn]);
-
+  const { isLoggedIn } = useUserStore();
 
   // PocketBase에서 북마크 데이터를 불러오는 함수
   useEffect(() => {
     const fetchBookmarks = async () => {
       try {
-        if (!user || !user.id) {
-          console.warn('User is not authenticated.');
-          return;
-        }
-
-        const userid = user.id;
+        const userid = JSON.parse(localStorage.getItem('pb_auth')).token.id;
         const bookmarks = await pb.collection('bookmarkItem').getFullList({
           filter: `user = "${userid}"`,
           sort: 'saveTime',
@@ -141,13 +122,8 @@ const CalendarPage = () => {
       }
     };
 
-    if (isLoggedIn) {
-      fetchBookmarks();
-    } else {
-      setBookmarkList([]); // Clear bookmark list if not logged in
-      setBookmarkMap({});
-    }
-  }, [isLoggedIn, user]);
+    fetchBookmarks();
+  }, []);
 
   // 이전/다음 북마크로 이동하는 함수
   const goToBookmark = (direction) => {
