--- src/stores/index.js
@@ -1,4 +1,0 @@
-// src/stores/index.js
-export { default as useLikeStore } from './likeStore';
-export { default as useUserStore } from './userStore';
-export { default as useWeatherStore } from './weatherStore';
